+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 
 - [Changelog](#changelog)
-  - [Next](#next)
+  - [2017.11.02.0](#201711020)
   - [2017.10.04.0](#201710040)
   - [2017.08.11.0](#201708110)
   - [2017.08.08.0](#201708080)
@@ -26,8 +26,14 @@
 
 ## Next
 
-* Added `Node.send_file` to allow files to be copied to nodes.
 * Backwards incompatible change: Rename `DCOS_Docker` backend to `Docker` backend.
+
+## 2017.11.02.0
+
+* Added `Node.send_file` to allow files to be copied to nodes.
+* Added `custom_master_mounts` to the DC/OS Docker backend.
+* Backwards incompatible change: Removed `files_to_copy_to_masters`.
+  Instead, use `custom_master_mounts` or `Node.send_file`.
 
 ## 2017.10.04.0